requirements.txt
python-mpd2==0.5.1
toml==0.7.0
git+git://github.com/quick2wire/quick2wire-python-api.git@0.0.0.2
